Added support for help command
Help command is used to get help on symbols from GNU APL.
If it is available remotely(i.e. remote protocol version
is > 1.5), the help will be requested and the contents
of gnu-apl--symbol-doc variable will be updated.

Note that the remote help command should return a list
of S-expressions which could be easily 'read-from-string'
and processed.

Also removed duplicated gnu-apl function.